Reorganization, division between Scrapy and Selenium
